ppo agent working with default args

+++ src/ppo/utils.py
@@ -233,7 +233,7 @@ def get_obs_preprocessor(obs_space):
     # handle cases where obs space is instance of gym.spaces.Box, gym.spaces.Dict, gym.spaces
 
     if isinstance(obs_space, gym.spaces.Box):
-        return lambda x: np.array(x).astype(np.float64)
+        return lambda x: np.array(x).astype(np.float32)
 
     elif isinstance(obs_space, gym.spaces.Dict):
         obs_space = obs_space.spaces
@@ -243,5 +243,5 @@ def get_obs_preprocessor(obs_space):
 def preprocess_images(images, device=None):
     # Bug of Pytorch: very slow if not first converted to numpy array
     images = np.array(images)
-    images = images.astype(np.float64)
+    images = images.astype(np.float32)
     return images
